Migrate get_dummies.
